New style mutation for 5 more ecommerce mutations
