+++ src/data/exchange.graphql
@@ -9,8 +9,9 @@ input ApproveOrderInput {
 type ApproveOrderPayload {
   # A unique identifier for the client performing the mutation.
   clientMutationId: String
-  errors: [String!]!
-  order: Order
+
+  # A union of success/failure
+  orderOrError: OrderOrFailureUnion!
 }
 
 # Autogenerated input type of CreateOrderWithArtwork
@@ -55,8 +56,9 @@ input FulfillAtOnceInput {
 type FulfillAtOncePayload {
   # A unique identifier for the client performing the mutation.
   clientMutationId: String
-  errors: [String!]!
-  order: Order
+
+  # A union of success/failure
+  orderOrError: OrderOrFailureUnion!
 }
 
 # A Fulfillment for an order
@@ -335,8 +337,9 @@ input RejectOrderInput {
 type RejectOrderPayload {
   # A unique identifier for the client performing the mutation.
   clientMutationId: String
-  errors: [String!]!
-  order: Order
+
+  # A union of success/failure
+  orderOrError: OrderOrFailureUnion!
 }
 
 # Represents either a shipping information or pickup
@@ -354,8 +357,9 @@ input SetPaymentInput {
 type SetPaymentPayload {
   # A unique identifier for the client performing the mutation.
   clientMutationId: String
-  errors: [String!]!
-  order: Order
+
+  # A union of success/failure
+  orderOrError: OrderOrFailureUnion!
 }
 
 # Autogenerated input type of SetShipping
@@ -371,8 +375,9 @@ input SetShippingInput {
 type SetShippingPayload {
   # A unique identifier for the client performing the mutation.
   clientMutationId: String
-  errors: [String!]!
-  order: Order
+
+  # A union of success/failure
+  orderOrError: OrderOrFailureUnion!
 }
 
 type Ship {
