+++ src/schema/__tests__/ecommerce/set_order_payment_mutation.test.js
@@ -4,7 +4,7 @@ import sampleOrder from "test/fixtures/results/sample_order"
 import exchangeOrderJSON from "test/fixtures/exchange/order.json"
 import { mockxchange } from "test/fixtures/exchange/mockxchange"
 import gql from "lib/gql"
-import { OrderSellerFields } from "./order_fields"
+import { OrderBuyerFields } from "./order_fields"
 
 let rootValue
 
@@ -13,8 +13,7 @@ describe("Approve Order Mutation", () => {
     const resolvers = {
       Mutation: {
         setPayment: () => ({
-          order: exchangeOrderJSON,
-          errors: [],
+          orderOrError: { order: exchangeOrderJSON },
         }),
       },
     }
@@ -28,18 +27,24 @@ describe("Approve Order Mutation", () => {
             orderId: "111",
             creditCardId: "1231-1232-4343-4343"
           }) {
-            result {
-              order {
-                ${OrderSellerFields}
+            orderOrError {
+              ... on OrderWithMutationSuccess {
+                order {
+                  ${OrderBuyerFields}
+                }
               }
-            errors
+              ... on OrderWithMutationFailure {
+                error {
+                  description
+                }
+              }
             }
           }
         }
     `
 
     return runQuery(mutation, rootValue).then(data => {
-      expect(data.setOrderPayment.result.order).toEqual(
+      expect(data.setOrderPayment.orderOrError.order).toEqual(
         sampleOrder(true, false)
       )
     })
