+++ src/schema/ecommerce/set_order_shipping_mutation.ts
@@ -13,6 +13,7 @@ import {
   RequestedFulfillmentFragment,
   BuyerSellerFields,
 } from "./query_helpers"
+import { OrderOrFailureUnionType } from "./types/order_or_error_union"
 
 const ShippingInputField = new GraphQLInputObjectType({
   name: "ShippingInputField",
@@ -71,9 +72,8 @@ export const SetOrderShippingMutation = mutationWithClientMutationId({
   description: "Sets shipping information for an order",
   inputFields: SetOrderShippingInput.getFields(),
   outputFields: {
-    result: {
-      type: OrderReturnType,
-      resolve: order => order,
+    orderOrError: {
+      type: OrderOrFailureUnionType,
     },
   },
   mutateAndGetPayload: (
@@ -98,37 +98,45 @@ export const SetOrderShippingMutation = mutationWithClientMutationId({
             shipping: $shipping
           }
         ) {
-          order {
-            id
-            code
-            currencyCode
-            state
-            ${BuyerSellerFields}
-            ${RequestedFulfillmentFragment}
-            itemsTotalCents
-            shippingTotalCents
-            taxTotalCents
-            commissionFeeCents
-            transactionFeeCents
-            buyerTotalCents
-            sellerTotalCents
-            updatedAt
-            createdAt
-            stateUpdatedAt
-            stateExpiresAt
-            lineItems {
-              edges {
-                node {
-                  id
-                  priceCents
-                  artworkId
-                  editionSetId
-                  quantity
+          orderOrError {
+            ... on EcommerceOrderWithMutationSuccess {
+              order {
+                id
+                code
+                currencyCode
+                state
+                ${BuyerSellerFields}
+                ${RequestedFulfillmentFragment}
+                itemsTotalCents
+                shippingTotalCents
+                taxTotalCents
+                commissionFeeCents
+                transactionFeeCents
+                buyerTotalCents
+                sellerTotalCents
+                updatedAt
+                createdAt
+                stateUpdatedAt
+                stateExpiresAt
+                lineItems {
+                  edges {
+                    node {
+                      id
+                      priceCents
+                      artworkId
+                      editionSetId
+                      quantity
+                    }
+                  }
                 }
               }
             }
+            ... on EcommerceOrderWithMutationFailure {
+              error {
+                description
+              }
+            }
           }
-          errors
         }
       }
     `
@@ -136,15 +144,6 @@ export const SetOrderShippingMutation = mutationWithClientMutationId({
       orderId,
       fulfillmentType,
       shipping,
-    }).then(result => {
-      if (result.errors) {
-        throw Error(result.errors.map(d => d.message))
-      }
-      const { order, errors } = result.data.ecommerce_setShipping
-      return {
-        order,
-        errors,
-      }
-    })
+    }).then(result => result.data!.ecommerce_setShipping)
   },
 })
